Fix: Replace deprecated ske service state endpoints with new ones
