+++ internal/pkg/services/ske/utils/utils_test.go
@@ -57,72 +57,6 @@ func (m *skeClientMocked) ListProviderOptionsExecute(_ context.Context) (*ske.Pr
 	return m.listProviderOptionsResp, nil
 }
 
-func TestProjectEnabled(t *testing.T) {
-	tests := []struct {
-		description     string
-		serviceDisabled bool
-		getProjectFails bool
-		getProjectResp  *ske.ProjectResponse
-		isValid         bool
-		expectedOutput  bool
-	}{
-		{
-			description:    "project enabled",
-			getProjectResp: &ske.ProjectResponse{State: ske.PROJECTSTATE_CREATED.Ptr()},
-			isValid:        true,
-			expectedOutput: true,
-		},
-		{
-			description:     "project disabled (404)",
-			serviceDisabled: true,
-			isValid:         true,
-			expectedOutput:  false,
-		},
-		{
-			description:    "project disabled 1",
-			getProjectResp: &ske.ProjectResponse{State: ske.PROJECTSTATE_CREATING.Ptr()},
-			isValid:        true,
-			expectedOutput: false,
-		},
-		{
-			description:    "project disabled 2",
-			getProjectResp: &ske.ProjectResponse{State: ske.PROJECTSTATE_DELETING.Ptr()},
-			isValid:        true,
-			expectedOutput: false,
-		},
-		{
-			description:     "get clusters fails",
-			getProjectFails: true,
-			isValid:         false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.description, func(t *testing.T) {
-			client := &skeClientMocked{
-				serviceDisabled:       tt.serviceDisabled,
-				getServiceStatusFails: tt.getProjectFails,
-				getServiceStatusResp:  tt.getProjectResp,
-			}
-
-			output, err := ProjectEnabled(context.Background(), client, testProjectId)
-
-			if tt.isValid && err != nil {
-				t.Errorf("failed on valid input")
-			}
-			if !tt.isValid && err == nil {
-				t.Errorf("did not fail on invalid input")
-			}
-			if !tt.isValid {
-				return
-			}
-			if output != tt.expectedOutput {
-				t.Errorf("expected output to be %t, got %t", tt.expectedOutput, output)
-			}
-		})
-	}
-}
-
 func TestClusterExists(t *testing.T) {
 	tests := []struct {
 		description      string
